classifier/requirements.txt
flask 
torch 
torchvision 
timm 
pillow